Renamed all instances of text_2_cypher text2cypher
